templates/twitter/requirements.txt
requests
gspread
google-auth
schedule
python-dotenv
